Rename ophanComponentEventTracking to just ophan
