+++ support-frontend/assets/components/subscriptionCheckouts/thankYou/marketingConsentContainer.jsx
@@ -4,7 +4,7 @@
 
 import { connect } from 'react-redux';
 import MarketingConsent from 'components/marketingConsentNew/marketingConsent';
-import { trackComponentClick } from 'helpers/tracking/ophanComponentEventTracking';
+import { trackComponentClick } from 'helpers/tracking/ophan';
 import type { Dispatch } from 'redux';
 import type { Action } from 'helpers/user/userActions';
 import type { Csrf as CsrfState } from 'helpers/csrf/csrfReducer';
