+++ support-frontend/assets/helpers/checkoutForm/onFormSubmitOld.js
@@ -2,7 +2,7 @@
 import type { ContributionType } from 'helpers/contributions';
 import { canContributeWithoutSigningIn, type UserTypeFromIdentityResponse } from 'helpers/identityApis';
 import { formElementIsValid, invalidReason } from 'helpers/checkoutForm/checkoutForm';
-import { trackCheckoutSubmitAttempt } from 'helpers/tracking/ophanComponentEventTracking';
+import { trackCheckoutSubmitAttempt } from 'helpers/tracking/ophan';
 import type { PaymentMethod } from 'helpers/paymentMethods';
 
 type OldFlowOrNewFlow = 'opf' | 'npf';
