+++ support-frontend/assets/pages/contributions-landing/contributionsLanding.jsx
@@ -2,7 +2,7 @@
 
 // ----- Imports ----- //
 
-import { trackPolyfillScriptStatus } from 'helpers/tracking/ophanComponentEventTracking';
+import { trackPolyfillScriptStatus } from 'helpers/tracking/ophan';
 import React from 'react';
 import { Provider } from 'react-redux';
 import { Route, BrowserRouter } from 'react-router-dom';
@@ -27,7 +27,7 @@ import ContributionThankYouContainer from './components/ContributionThankYou/Con
 import { setUserStateActions } from './setUserStateActions';
 import ConsentBanner from '../../components/consentBanner/consentBanner';
 import './contributionsLanding.scss';
-import { trackPolyfilledObjectFunctions } from '../../helpers/tracking/ophanComponentEventTracking';
+import { trackPolyfilledObjectFunctions } from '../../helpers/tracking/ophan';
 
 const polyfillSuccess = window.guardian.polyfillScriptStatus;
 trackPolyfillScriptStatus(polyfillSuccess);
